Revert "Move test fixture"
This reverts commit 059d640dd31304642940deb28237a93367ac5e2e.
