+++ app/tests/web/db/test_crud.py
@@ -10,10 +10,92 @@ from app.shared.settings import Settings
 from app.web.db import crud
 
 
+authors = ["rick@example.com", "morty@example.com", "jerry@example.com"]
+
+
+@pytest.fixture()
+def test_data(db_session):
+    # creates 3 users
+    for email in authors:
+        db_session.add(models.User(email=email))
+    db_session.commit()
+    assert db_session.query(models.User).count() == 3
+
+    # creates 100 archives for 3 users over 2 months with repeating URLs
+    for i in range(100):
+        author = authors[i % 3]
+        archive = models.Archive(
+            id=f"archive-id-456-{i}",
+            url=f"https://example-{i % 3}.com",
+            result={},
+            public=author == "jerry@example.com",
+            author_id=author,
+            group_id="spaceship"
+            if author == "morty@example.com" and i % 2 == 0
+            else None,
+            created_at=datetime(2021, (i % 2) + 1, (i % 25) + 1),
+        )
+        if i % 5 == 0:
+            archive.tags.append(models.Tag(id=f"tag-{i}"))
+        if i % 10 == 0:
+            archive.tags.append(models.Tag(id=f"tag-second-{i}"))
+        if i % 4 == 0:
+            archive.tags.append(models.Tag(id=f"tag-third-{i}"))
+        for j in range(10):
+            archive.urls.append(
+                models.ArchiveUrl(
+                    url=f"https://example-{i}.com/{j}", key=f"media_{j}"
+                )
+            )
+        db_session.add(archive)
+
+    # creates a sheet for each user
+    for i, email in enumerate(authors):
+        db_session.add(
+            models.Sheet(
+                id=f"sheet-{i}",
+                name=f"sheet-{i}",
+                author_id=email,
+                group_id=None,
+                frequency="daily",
+            )
+        )
+        if email == "rick@example.com":
+            db_session.add(
+                models.Sheet(
+                    id=f"sheet-{i}-2",
+                    name=f"sheet-{i}-2",
+                    author_id=email,
+                    group_id="spaceship",
+                    frequency="hourly",
+                )
+            )
+
+    db_session.commit()
+
+    assert db_session.query(models.Archive).count() == 100
+    assert db_session.query(models.Tag).count() == 20 + 10 + 25
+    assert db_session.query(models.ArchiveUrl).count() == 1000
+    assert (
+        db_session.query(models.ArchiveUrl)
+        .filter(models.ArchiveUrl.archive_id == "archive-id-456-0")
+        .count()
+        == 10
+    )
+
+    # setup groups
+    assert db_session.query(models.Group).count() == 0
+    from app.web.db import crud
+
+    crud.upsert_user_groups(db_session)
+    assert db_session.query(models.Group).count() == 4
+    assert db_session.query(models.User).count() == 3
+
+
 def test_search_archives_by_url(test_data, db_session):
     from app.web.config import ALLOW_ANY_EMAIL
 
-    # Rick's archives are private
+    # rick's archives are private
     assert (
         len(
             crud.search_archives_by_url(
